requirements.txt
django==3.0
mssql-django==1.0b1
django-environ==0.4.5
django-filter==2.4.0
djangorestframework-simlejwt==4.7.1
djangorestframework==3.12.4